requirements.txt
streamlit
scrapegraphai
duckdb
pandas
pyarrow
openai
python-dotenv
agno
phi
playwright
